Abstract configuration into enum
