+++ Sources/SwiftNetKit/Models/MethodType.swift
@@ -1,0 +1,14 @@
+//
+//  MethodType.swift
+//
+//
+//  Created by Sam Gilmore on 7/17/24.
+//
+
+enum MethodType: String {
+    case get = "GET"
+    case post = "POST"
+    case delete = "DELETE"
+    case put = "PUT"
+    case patch = "PATCH"
+}
