+++ Sources/SwiftNetKit/Models/NetworkError.swift
@@ -5,8 +5,6 @@
 //  Created by Sam Gilmore on 7/16/24.
 //
 
-import Foundation
-
 public enum NetworkError: Error {
     case invalidResponse
     case decodingFailed
