+++ Sources/SwiftNetKit/Models/SessionConfiguration.swift
@@ -1,0 +1,12 @@
+//
+//  SessionConfiguration.swift
+//
+//
+//  Created by Sam Gilmore on 7/17/24.
+//
+
+public enum SessionConfiguration {
+    case `default`
+    case ephemeral
+    case background(String)
+}
